requirements.txt
matplotlib==3.6.2
numpy==1.23.4
pandas==1.5.1
albumentations==1.3.0
timm==0.6.11
wandb==0.13.5
torchinfo==1.7.1
torch==1.13.0
torchvision==0.14.0
torchaudio==0.13.0
grad-cam==1.4.6
black==21.9b0
pylint==2.7.4
pytorch-lightning==1.7.7
click==7.1.2
imagesize==1.4.1
torchmetrics==0.10.3
pydicom==2.3.1
pydantic==1.10.2
tabulate==0.9.0
hydra-core==1.3.1
omegaconf==2.3.0